Refactor export spot counts per timepoint
